requirements.txt
django==4.1.2
djangorestframework==3.14.0 # drf toolkit - easy to use rest api framework
django-cors-headers # adds Cross-Origin-Resourse-Sharing (CORS), allowing in-browser requests from other domains
# django-extensions
psycopg2-binary==2.9.4  # postgresql database adapter for python
python-dotenv>=0.20.0 # python-dotenv reads key-value pairs from a .env file and can set them as environment variables.
djangorestframework-simplejwt==5.2.1 # JSON Web Token authentication for Django REST Framework
django-cryptography>=1.1.0 # django-cryptography is a Django app that provides a set of fields for storing encrypted data in your models.
Pillow  # Pillow is the friendly PIL fork by Alex Clark and Contributors. PIL is the Python Imaging Library by Fredrik Lundh and Contributors.

# django-filter TODO: add that to the project

django-guardian~=2.4.0